basic connection with docker monog
