+++ backend/api/database.py
@@ -1,5 +1,4 @@
 from sqlalchemy import create_engine
-from .models import Base
 
 from sqlalchemy.orm import sessionmaker
 
@@ -9,6 +8,3 @@ engine = create_engine(
     SQLALCHEMY_DATABASE_URL, connect_args={"check_same_thread": False}
 )
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-
-Base.metadata.create_all(bind=engine)
