+++ backend/api/mongo.py
@@ -1,14 +1,5 @@
 from pymongo import MongoClient
-from models import Applicant, Admin, Building
+from .models import Applicant, Admin, Building
 
-client = MongoClient("mongodb://admin:pass@localhost:27017")
+client = MongoClient("mongodb://root:examplepassword@localhost:27017/?authSource=admin")
 db = client.matching
-
-
-def get_db():
-    db = client.matching
-    try:
-        yield db
-    finally:
-        db.close()
-
