+++ backend/api/routes/admin.py
@@ -1,16 +1,14 @@
-from fastapi import APIRouter, HTTPException, Depends
-from sqlalchemy.orm import Session, joinedload
-from ..utlils import get_password_hash, verify_password
-from ..utlils import get_db
-from ..models import Admin, Building, Applicant, AdminRanking
+from fastapi import APIRouter, HTTPException
+from sqlalchemy.orm import joinedload
+from ..utlils import get_password_hash, verify_password, db
+from ..models import Admin, Building, Applicant
 from ..schemas import AdminLogin, AdminRankingCreate
-
 
 router = APIRouter()
 
 
 @router.post("/admin_login/")
-def admin_login(admin: AdminLogin, db: Session = Depends(get_db)):
+def admin_login(admin: AdminLogin):
     db_admin = db.query(Admin).filter(Admin.du_id == admin.du_id).first()
     if not db_admin:
         raise HTTPException(status_code=400, detail="Invalid credentials")
@@ -22,7 +20,7 @@ def admin_login(admin: AdminLogin, db: Session = Depends(get_db)):
 
 
 @router.get("/admin/{admin_id}")
-def get_admin(admin_id: str, db: Session = Depends(get_db)):
+def get_admin(admin_id: str):
     admin_db = (
         db.query(Admin)
         .options(joinedload(Admin.buildings))
@@ -36,7 +34,7 @@ def get_admin(admin_id: str, db: Session = Depends(get_db)):
 
 
 @router.get("/admin_given_preferences/{admin_id}")
-def get_admin_given_preferences(admin_id: str, db: Session = Depends(get_db)):
+def get_admin_given_preferences(admin_id: str):
     admin_db = db.query(Admin).filter(Admin.id == admin_id).first()
     if not admin_db:
         raise HTTPException(status_code=404, detail="Admin not found")
@@ -45,8 +43,8 @@ def get_admin_given_preferences(admin_id: str, db: Session = Depends(get_db)):
 
 
 @router.get("/get_all_admins/")
-def get_all_admins(db: Session = Depends(get_db)):
-    admins = db.query(Admin).options(joinedload(Admin.building)).all()
+def get_all_admins():
+    admins = db.query(Admin).options(joinedload(Admin.buildings)).all()
     if not admins:
         raise HTTPException(status_code=404, detail="No admins found")
 
@@ -54,7 +52,7 @@ def get_all_admins(db: Session = Depends(get_db)):
 
 
 @router.post("/admin_rank/")
-def admin_rank(data: AdminRankingCreate, db: Session = Depends(get_db)):
+def admin_rank(data: AdminRankingCreate):
     admin = db.query(Admin).filter(Admin.id == data.admin_id).first()
     if not admin:
         raise HTTPException(status_code=404, detail="Admin not found")
@@ -74,17 +72,15 @@ def admin_rank(data: AdminRankingCreate, db: Session = Depends(get_db)):
         db.add(new_ranking)
 
     admin.given_preferences = True
-
     db.commit()
     return {"message": "Rankings submitted successfully!"}
 
 
-@router.get("/admin_rankings_by_admin/{admin_du_id}")
-def view_admin_rankings(admin_id: int, db: Session = Depends(get_db)):
+@router.get("/admin_rankings_by_admin/{admin_id}")
+def view_admin_rankings(admin_id: int):
     admin = db.query(Admin).filter(Admin.id == admin_id).first()
     if not admin:
         raise HTTPException(status_code=404, detail="Admin not found")
 
     rankings = db.query(AdminRanking).filter(AdminRanking.admin_id == admin_id).all()
-
     return rankings
