+++ backend/api/routes/applicant.py
@@ -3,7 +3,7 @@ from fastapi import APIRouter, HTTPException, Depends, UploadFile, File
 from ..schemas import UserCreate, RAAppCreate, UserRead, UserLogin
 from ..models import Applicant  # your Pydantic/ORM model used for validating inserts
 from ..utlils import get_password_hash, verify_password
-from ..mongo import get_db
+from ..mongo import db
 from bson import ObjectId
 import os
 
@@ -11,7 +11,9 @@ router = APIRouter()
 
 
 @router.post("/create_applicant/")
-def create_user(user: UserCreate, db=Depends(get_db)):
+def create_user(
+    user: UserCreate,
+):
     applicants = db["applicants"]
     existing = applicants.find_one({"du_id": user.du_id})
     if existing:
@@ -28,13 +30,16 @@ def create_user(user: UserCreate, db=Depends(get_db)):
         email=user.email,
         password=get_password_hash(user.password),
         year_in_college=user.year_in_college,
+        resume_path=None,
+        pref=None,
+        is_returner=None,
     )
     applicants.insert_one(new_user.model_dump())
     return {"message": "User created successfully!", "id": new_user.applicant_id}
 
 
 @router.post("/login/")
-def login(user: UserLogin, db=Depends(get_db)):
+def login(user: UserLogin):
     applicants = db["applicants"]
     db_user = applicants.find_one({"du_id": user.du_id})
     if not db_user or not verify_password(user.password, db_user["password"]):
@@ -43,7 +48,7 @@ def login(user: UserLogin, db=Depends(get_db)):
 
 
 @router.post("/apply/")
-def apply(data: RAAppCreate, db=Depends(get_db)):
+def apply(data: RAAppCreate):
     applicants = db["applicants"]
     user = applicants.find_one({"applicant_id": data.id})
     if not user:
@@ -64,7 +69,7 @@ def apply(data: RAAppCreate, db=Depends(get_db)):
 
 
 @router.post("/upload_resume/{id}")
-def upload_resume(id: int, resume: UploadFile = File(...), db=Depends(get_db)):
+def upload_resume(id: int, resume: UploadFile = File(...)):
     applicants = db["applicants"]
     user = applicants.find_one({"applicant_id": id})
     if not user:
@@ -82,7 +87,7 @@ def upload_resume(id: int, resume: UploadFile = File(...), db=Depends(get_db)):
 
 
 @router.get("/applicants/{du_id}")
-def get_applicant(du_id: str, db=Depends(get_db)):
+def get_applicant(du_id: str, db):
     applicants = db["applicants"]
     doc = applicants.find_one({"du_id": du_id})
     if not doc:
@@ -91,7 +96,7 @@ def get_applicant(du_id: str, db=Depends(get_db)):
 
 
 @router.get("/get_all_applicants/")
-def all_applicants_with_preferences(db=Depends(get_db)):
+def all_applicants_with_preferences():
     applicants = db["applicants"]
     docs = list(applicants.find({}))
     if not docs:
@@ -100,7 +105,7 @@ def all_applicants_with_preferences(db=Depends(get_db)):
 
 
 @router.get("/applicant_given_preferences/{applicant_id}")
-def get_applicant_given_preferences(applicant_id: int, db=Depends(get_db)):
+def get_applicant_given_preferences(applicant_id: int):
     applicants = db["applicants"]
     doc = applicants.find_one({"applicant_id": applicant_id})
     if not doc:
