+++ backend/api/routes/building.py
@@ -1,13 +1,13 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
-from ..utlils import get_db
+from ..utlils import db
 from ..models import Building
 
 router = APIRouter()
 
 
 @router.get("/buildings/")
-def get_buildings(db: Session = Depends(get_db)):
+def get_buildings():
     buildings = db["applicants"]
     list_of_buildings = buildings.find_one({})
     if not list_of_buildings:
