+++ backend/requirements.txt
@@ -36,3 +36,4 @@ uvicorn==0.34.2
 uvloop==0.21.0
 watchfiles==1.0.5
 websockets==15.0.1
+pymongo
